+++ pkg/mesh_install/installer.go
@@ -1,16 +1,18 @@
-// Package installer exposes functions for applying resources to a Kubernetes cluster.
+// Package mesh_install exposes functions for applying resources to a Kubernetes cluster.
 // Its exposed functions receive a k8sClient for communicating with the cluster.
-package installer
+package mesh_install
 
 import (
 	"context"
+	"github.com/cloudflare/cfssl/csr"
 	configv1 "github.com/openshift/api/config/v1"
+	"strings"
 	"time"
 
 	"github.com/greymatter-io/operator/api/v1alpha1"
 	"github.com/greymatter-io/operator/pkg/cfsslsrv"
-	"github.com/greymatter-io/operator/pkg/cli"
 	"github.com/greymatter-io/operator/pkg/cuemodule"
+	"github.com/greymatter-io/operator/pkg/gmapi"
 	"github.com/greymatter-io/operator/pkg/k8sapi"
 
 	corev1 "k8s.io/api/core/v1"
@@ -22,13 +24,13 @@ import (
 )
 
 var (
-	logger = ctrl.Log.WithName("installer")
+	logger = ctrl.Log.WithName("mesh_install")
 )
 
 // Installer stores a map of version.Version and a distinct version.Sidecar for each mesh.
 type Installer struct {
-	*cli.CLI  // Grey Matter CLI
-	k8sClient *client.Client
+	*gmapi.CLI // Grey Matter CLI
+	k8sClient  *client.Client
 
 	cfssl *cfsslsrv.CFSSLServer
 
@@ -50,26 +52,25 @@ type Installer struct {
 	// Root on disk of the operator CUE. Used for reloading the default configs on teardown
 	CueRoot string
 
-	// The cluster ingress domain
-	clusterIngressDomain string
+	// Operator config loadable from CUE
+	Config cuemodule.Config
 
-	// Flags loadable from CUE
-	Flags cuemodule.Flags
+	// Looked up on start
+	clusterIngressDomain string
 }
 
 // New returns a new *Installer instance for installing Grey Matter components and dependencies.
-func New(c *client.Client, operatorCUE *cuemodule.OperatorCUE, initialMesh *v1alpha1.Mesh, cueRoot string, gmcli *cli.CLI, cfssl *cfsslsrv.CFSSLServer, clusterIngressName string) (*Installer, error) {
-	flags, defaults := operatorCUE.ExtractFlagsAndDefaults()
+func New(c *client.Client, operatorCUE *cuemodule.OperatorCUE, initialMesh *v1alpha1.Mesh, cueRoot string, gmcli *gmapi.CLI, cfssl *cfsslsrv.CFSSLServer) (*Installer, error) {
+	config, defaults := operatorCUE.ExtractFlagsAndDefaults()
 	gmcli.SidecarList = defaults.RedisSpireSubjects
 	return &Installer{
-		CLI:                gmcli,
-		k8sClient:          c,
-		cfssl:              cfssl,
-		clusterIngressName: clusterIngressName,
-		OperatorCUE:        operatorCUE,
-		Mesh:               initialMesh,
-		CueRoot:            cueRoot,
-		Flags:              flags,
+		CLI:         gmcli,
+		k8sClient:   c,
+		cfssl:       cfssl,
+		OperatorCUE: operatorCUE,
+		Mesh:        initialMesh,
+		CueRoot:     cueRoot,
+		Config:      config,
 	}, nil
 }
 
@@ -102,7 +103,7 @@ func (i *Installer) Start(ctx context.Context) error {
 	}
 
 	// TODO the operator itself can launch _all_ the Spire manifests, as before, after a check that they aren't already there.
-	if i.Flags.Spire {
+	if i.Config.Spire {
 		spireSecret, err = injectPKI(spireSecret, i.cfssl)
 		if err != nil {
 			logger.Error(err, "Error while attempting to apply spire server-ca secret", "secret object", spireSecret)
@@ -111,9 +112,8 @@ func (i *Installer) Start(ctx context.Context) error {
 		k8sapi.Apply(i.k8sClient, spireSecret, i.owner, k8sapi.CreateOrUpdate)
 	}
 
-	// TODO bring this back when you get to re-enabling (testing with) OpenShift
 	// Try to get the OpenShift cluster ingress domain if it exists.
-	clusterIngressDomain, ok := getOpenshiftClusterIngressDomain(i.k8sClient, i.clusterIngressName)
+	clusterIngressDomain, ok := getOpenshiftClusterIngressDomain(i.k8sClient, i.Config.ClusterIngressName)
 	if ok {
 		// TODO: When not in OpenShift, check for other supported ingress class types such as Nginx or Voyager.
 		// If no supported ingress types are found, just assume the user will configure ingress on their own.
@@ -123,7 +123,7 @@ func (i *Installer) Start(ctx context.Context) error {
 
 	// Immediately apply the default mesh from the CUE if the flag is set
 	go func() {
-		if i.Flags.AutoApplyMesh {
+		if i.Config.AutoApplyMesh {
 			time.Sleep(100 * time.Second) // DEBUG - does this work if we wait long enough for it to register the webhooks?
 			k8sapi.Apply(i.k8sClient, i.Mesh, nil, k8sapi.CreateOrUpdate)
 		}
@@ -163,7 +163,6 @@ func getImagePullSecret(c *client.Client) *corev1.Secret {
 	}
 }
 
-// TODO bring this back when we re-enable OpenShift support
 func getOpenshiftClusterIngressDomain(c *client.Client, ingressName string) (string, bool) {
 	clusterIngressList := &configv1.IngressList{}
 	if err := (*c).List(context.TODO(), clusterIngressList); err != nil {
@@ -193,4 +192,26 @@ func isSupportedKubernetesIngressClassPresent(c client.Client) bool {
 		}
 	}
 	return false
+}
+
+func injectPKI(secret *corev1.Secret, cs *cfsslsrv.CFSSLServer) (*corev1.Secret, error) {
+	root := cs.GetRootCA()
+	ca, caKey, err := cs.RequestIntermediateCA(csr.CertificateRequest{
+		CN:         "Grey Matter SPIFFE Intermediate CA",
+		KeyRequest: &csr.KeyRequest{A: "ecdsa", S: 256},
+		Names: []csr.Name{
+			{C: "US", ST: "VA", L: "Alexandria", O: "Grey Matter"},
+		},
+	})
+	if err != nil {
+		return nil, err
+	}
+
+	secret.StringData = map[string]string{
+		"root.crt":         string(root),
+		"intermediate.crt": strings.Join([]string{string(ca), string(root)}, "\n"),
+		"intermediate.key": string(caKey),
+	}
+
+	return secret, nil
 }